Add detection of CgroupSetup to systemd.Manager
This detects whether systemd is using cgroupv1, cgroupv2 or both.

Detection is performed according to the algorithm describe in:
https://systemd.io/CGROUP_DELEGATION.html#three-different-tree-setups-

The detected CgroupSetup is then stored in the systemd.Manager object.

Tested by running a binary using this code in three setups using each of
the combinations and confirming detection worked as expected.

Signed-off-by: Filipe Brandenburger <filbranden@google.com>