Update getSubsystemPath to work on unified cgroup setup
Always use the hardcoded cgroupRoot as path, since that's the only
configuration supported in systemd anyways.

Setting subsystem to the empty string makes getControllerPath() work,
since the contents of /proc/$PID/cgroup will have an empty string on
the second field for the cgroupv2 mount.

With this commit in place, using runc on unified cgroup stopped failing
with this message:

  process_linux.go:275: applying cgroup configuration for process caused "mountpoint for devices not found"

And started failing with this message:

  process_linux.go:275: applying cgroup configuration for process caused "open /sys/fs/cgroup/machine.slice/cpuset.cpus: no such file or directory"

Signed-off-by: Filipe Brandenburger <filbranden@google.com>